Use refactored Registers types
This change switches from Register/Registers from helix-core to the
new registers types introduced in the parent commits. This is an
unfortunately noisy change: we need to update virtually all callsites
that access the registers.

We need to move Registers out of the Editor type to make this work:
Registers taking '&mut self' and Editor parameters ('&' or '&mut')
causes a borrowing conflict if Registers is part of Editor. We move
Registers up to the Context types to resolve this.